Changed the way MatchView.tsk handles game finish and restart, also reorganized file structure
