+++ src/components/App.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from "react-router";
+
+
+function App() {
+  return(
+    <>
+      <h1>Tic-Tac-Toe</h1>
+      <Outlet />
+    </>
+  )
+}
+
+export default App;
